Android and iOS build locations are optional

+++ build-assets.py
@@ -20,20 +20,6 @@ parser.add_option("-c", action="store_true", dest="cleanup", help="clean up file
 
 (options, args) = parser.parse_args()
 
-if (options.ios == None or options.android == None):
-    print("Missing iOS or Android path, please see build-assets.py -h")
-    sys.exit(1)
-
-if (os.path.isdir(options.ios) == False or
-    os.path.exists(options.ios + '/NZSLDict/main.m') == False):
-    print("Invalid iOS path")
-    sys.exit(1)
-
-if (os.path.isdir(options.android) == False or
-    os.path.exists(options.android + '/build.gradle') == False):
-    print("Invalid Android path")
-    sys.exit(1)
-
 filename = 'dnzsl-xmldump.xml'
 
 print("Step 1: Fetching the latest signs from Freelex")
@@ -80,7 +66,7 @@ for path, dirs, files in os.walk("assets/"):
         print_run_msg(create_thumbnail_cmd)
         os.system(create_thumbnail_cmd)
 
-print("Step 6p: Shrink images for distribution")
+print("Step 6b: Shrink images for distribution")
 # In order to keep the app size small, we need to run a series
 # of compressions over the images
 
@@ -105,52 +91,55 @@ for path, dirs, files in os.walk("assets/"):
         print_run_msg(optipng_cmd)
         os.system(optipng_cmd)
 
-print("Step 7a: Update iOS app images")
 
-ios_asset_path = options.ios + "/Data/picture/"
+if options.ios is not None:
+    print("Step 7a: Update iOS app images")
 
-if os.path.isdir(ios_asset_path):
-    shutil.rmtree(ios_asset_path)
+    ios_asset_path = options.ios + "/Data/picture/"
 
-os.makedirs(ios_asset_path)
+    if os.path.isdir(ios_asset_path):
+        shutil.rmtree(ios_asset_path)
 
-# re-create the .gitkeep file in the assets dir (it is not necessary for the
-# app but keeps that dir in the git repo)
-os.system("touch " + ios_asset_path + ".gitkeep")
+    os.makedirs(ios_asset_path)
 
-for path, dirs, files in os.walk("assets/"):
-    for filename in files:
-        cp_cmd = "cp assets/" + filename + " " + ios_asset_path
-        print_run_msg(cp_cmd)
-        os.system(cp_cmd)
+    # re-create the .gitkeep file in the assets dir (it is not necessary for the
+    # app but keeps that dir in the git repo)
+    os.system("touch " + ios_asset_path + ".gitkeep")
 
-print("Step 7b: Update iOS app nzsl.db")
-os.system("cp nzsl.db " + options.ios + "/Data/")
+    for path, dirs, files in os.walk("assets/"):
+        for filename in files:
+            cp_cmd = "cp assets/" + filename + " " + ios_asset_path
+            print_run_msg(cp_cmd)
+            os.system(cp_cmd)
 
-print("Step 8a: Update Android app images")
-android_asset_path = options.android + "/app/src/main/assets/images/signs/"
+    print("Step 7b: Update iOS app nzsl.db")
+    os.system("cp nzsl.db " + options.ios + "/Data/")
 
-if os.path.isdir(android_asset_path):
-    shutil.rmtree(android_asset_path)
+if options.android is not None:
+    print("Step 8a: Update Android app images")
+    android_asset_path = options.android + "/app/src/main/assets/images/signs/"
 
-os.makedirs(android_asset_path)
+    if os.path.isdir(android_asset_path):
+        shutil.rmtree(android_asset_path)
 
-# re-create the .gitkeep file in the assets dir (it is not necessary for the
-# app but keeps that dir in the git repo)
-os.system("touch " + android_asset_path + ".gitkeep")
+    os.makedirs(android_asset_path)
 
-for path, dirs, files in os.walk("assets/"):
-    for filename in files:
-        cp_cmd = "cp assets/" + filename + " " + android_asset_path
-        print_run_msg(cp_cmd)
-        os.system(cp_cmd)
+    # re-create the .gitkeep file in the assets dir (it is not necessary for the
+    # app but keeps that dir in the git repo)
+    os.system("touch " + android_asset_path + ".gitkeep")
 
-print("Step 8b: Update Android app nzsl.dat")
-android_db_path = options.android + "/app/src/main/assets/db/"
-if not os.path.exists(android_db_path):
-    os.makedirs(android_db_path)
+    for path, dirs, files in os.walk("assets/"):
+        for filename in files:
+            cp_cmd = "cp assets/" + filename + " " + android_asset_path
+            print_run_msg(cp_cmd)
+            os.system(cp_cmd)
 
-os.system("cp nzsl.dat " + android_db_path)
+    print("Step 8b: Update Android app nzsl.dat")
+    android_db_path = options.android + "/app/src/main/assets/db/"
+    if not os.path.exists(android_db_path):
+        os.makedirs(android_db_path)
+
+    os.system("cp nzsl.dat " + android_db_path)
 
 if options.cleanup:
     print("Step 9: Cleanup")
